+++ avaliacao_2/ms_ranking.py
@@ -1,29 +1,32 @@
 import json
 import pika
+import os
 from Crypto.Signature import pkcs1_15
 from Crypto.Hash import SHA256
 from Crypto.PublicKey import RSA
 
 HOT_DEAL_THRESHOLD = 3
 
-key = RSA.generate(2048)
-
 pwd = b'senha'
-with open("./private_keys/msranking_privatekey.pem", "wb") as f:
-    data = key.export_key(passphrase=pwd,
-                          pkcs=8,
-                          protection='PBKDF2WithHMAC-SHA512AndAES256-CBC',
-                          prot_params={'iteration_count':131072})
-    f.write(data)
 
-with open("./private_keys/msranking_privatekey.pem", "rb") as f:
-    data = f.read()
-    mykey = RSA.import_key(data, pwd)
+if os.path.exists("./private_keys/msranking_privatekey.pem"):
+    with open("./private_keys/msranking_privatekey.pem", "rb") as f:
+        data = f.read()
+        mykey = RSA.import_key(data, pwd)
 
-with open("./public_keys/msranking_publickey.pem", "wb") as f:
-    public_key = key.publickey()
-    data = public_key.export_key()
-    f.write(data)
+else:
+    mykey = RSA.generate(2048)
+    with open("./private_keys/msranking_privatekey.pem", "wb") as f:
+        data = mykey.export_key(passphrase=pwd,
+                              pkcs=8,
+                              protection='PBKDF2WithHMAC-SHA512AndAES256-CBC',
+                              prot_params={'iteration_count':131072})
+        f.write(data)
+
+    with open("./public_keys/msranking_publickey.pem", "wb") as f:
+        public_key = mykey.publickey()
+        data = public_key.export_key()
+        f.write(data)
 
 lista_promocoes = {}
